+++ src/Ext/Techno/Body.Update.cpp
@@ -16,6 +16,7 @@
 #include <Ext/House/Body.h>
 #include <Ext/WeaponType/Body.h>
 #include <Ext/Scenario/Body.h>
+#include <Ext/Script/Body.h>
 #include <Utilities/EnumFunctions.h>
 #include <Utilities/AresFunctions.h>
 
@@ -42,7 +43,8 @@ void TechnoExt::ExtData::OnEarlyUpdate()
 	this->ApplyMindControlRangeLimit();
 	this->UpdateRecountBurst();
 	this->UpdateRearmInEMPState();
-
+	this->UpdateRandomTargets();
+	
 	if (this->AttackMoveFollowerTempCount)
 		this->AttackMoveFollowerTempCount--;
 }
@@ -2087,5 +2089,116 @@ void TechnoExt::ExtData::UpdateTintValues()
 	{
 		auto const pShieldType = this->Shield->GetType();
 		calculateTint(Drawing::RGB_To_Int(pShieldType->Tint_Color), static_cast<int>(pShieldType->Tint_Intensity * 1000), pShieldType->Tint_VisibleToHouses);
+	}
+}
+
+// Checking & cleanning RandomTargets data, if necessary
+void TechnoExt::ExtData::UpdateRandomTargets()
+{
+	auto pThis = this->OwnerObject();
+	
+	if (!pThis || pThis->RearmTimer.GetTimeLeft() > 0)// || !pThis->Target)
+		return;
+
+	bool hasWeapons = ScriptExt::IsUnitArmed(pThis);
+	if (!hasWeapons)
+		return;
+
+	auto pExt = this;
+	bool isBuilding = pThis->WhatAmI() != AbstractType::Building;
+	auto const pType = this->TypeExtData->OwnerObject();
+	int weaponIndex = pExt->OriginalTargetWeaponIndex;
+
+	if (weaponIndex < 0)
+		return;
+
+	// Distance & weapon checks
+	auto const pWeapon = pThis->GetWeapon(weaponIndex)->WeaponType;
+	if (!pWeapon)
+		return;
+
+	auto const pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
+	if (!pWeaponExt || pWeaponExt->RandomTarget <= 0.0)
+		return;
+
+	if (!IsValidTechno(pExt->CurrentRandomTarget))
+	{
+		pExt->ResetRandomTarget = false;
+		pExt->CurrentRandomTarget = nullptr;
+	}
+
+	// Make sure random targets are off if original target is destroyed or not exists
+	if (!IsValidTechno(pExt->OriginalTarget))
+	{
+		pExt->ResetRandomTarget = false;
+		pExt->CurrentRandomTarget = nullptr;
+		pExt->OriginalTarget = nullptr;
+
+		if (pExt->OriginalTargetWeaponIndex >= 0)
+			pThis->SetTarget(nullptr);
+
+		pExt->OriginalTargetWeaponIndex = -1;
+
+		if (pThis->SpawnManager)
+			pThis->SpawnManager->ResetTarget();
+
+		return;
+	}
+
+	// Check: force if the target must be reset or the object's target differs from the current random target
+	if (pExt->ResetRandomTarget || weaponIndex < 0 || (pExt->CurrentRandomTarget && pExt->CurrentRandomTarget != pThis->Target))
+	{
+		pExt->ResetRandomTarget = false;
+		pExt->CurrentRandomTarget = nullptr;
+		//pExt->OriginalTargetWeaponIndex = -1;
+		pThis->Target = pExt->OriginalTarget;
+
+		if (pThis->SpawnManager)
+			pThis->SpawnManager->ResetTarget();
+
+		return;
+	}
+
+	int minimumRange = pWeapon->MinimumRange;
+	int range = pWeapon->Range;
+	range += pExt->OriginalTarget->IsInAir() ? pType->AirRangeBonus : 0;
+	int distanceToOriginalTarget = pThis->DistanceFrom(pExt->OriginalTarget);
+	int distanceToCurrentRandomTarget = pExt->CurrentRandomTarget ? pThis->DistanceFrom(pExt->CurrentRandomTarget) : 0;
+	int isInvalidRangeForTheBuilding = distanceToCurrentRandomTarget > range || distanceToCurrentRandomTarget < minimumRange;
+
+	if (distanceToOriginalTarget < minimumRange
+		|| distanceToOriginalTarget > range
+		|| (isBuilding && pExt->CurrentRandomTarget) && isInvalidRangeForTheBuilding)
+	{
+		pExt->ResetRandomTarget = false;
+		pExt->CurrentRandomTarget = nullptr;
+		pThis->SetTarget(pExt->OriginalTarget);
+		pExt->OriginalTarget = nullptr;
+		pExt->OriginalTargetWeaponIndex = -1;
+		//pThis->SetTarget(nullptr);
+	}
+
+	if (pThis->SpawnManager)
+	{
+		if (!pExt->OriginalTarget)
+			pThis->SpawnManager->ResetTarget();
+
+		for (auto pSpawn : pThis->SpawnManager->SpawnedNodes)
+		{
+			if (!pSpawn->Unit || pSpawn->IsSpawnMissile)
+				continue;
+
+			auto pSpawnExt = TechnoExt::ExtMap.Find(pSpawn->Unit);
+			pSpawnExt->OriginalTarget = pExt->OriginalTarget;
+
+			if (!pSpawnExt->OriginalTarget)
+			{
+				pSpawnExt->ResetRandomTarget = false;
+				pSpawnExt->CurrentRandomTarget = nullptr;
+				pSpawnExt->OriginalTarget = nullptr;
+				pSpawnExt->OriginalTargetWeaponIndex = -1;
+				pSpawn->Unit->SetTarget(nullptr);
+			}
+		}
 	}
 }