+++ src/Ext/Techno/Hooks.cpp
@@ -15,6 +15,7 @@
 #include <Ext/WarheadType/Body.h>
 #include <Ext/WeaponType/Body.h>
 #include <Ext/TechnoType/Body.h>
+#include <Ext/Script/Body.h>
 #include <Utilities/EnumFunctions.h>
 #include <Utilities/Helpers.Alex.h>
 #include <Utilities/AresHelper.h>